[REF] stock, *: rename stock.quant.package
Rename the package model from `stock.quant.package` to `stock.package`
to better match the stock naming scheme.

Also moves the code relative to `stock.package` to their own files for
easier searches.

Task-4314600
