+++ addons/stock/data/stock_sequence_data.xml
@@ -43,10 +43,10 @@
             <field name="company_id" eval="False"/>
         </record>
 
-        <!-- Sequences for stock.quant.package -->
-        <record id="seq_quant_package" model="ir.sequence">
+        <!-- Sequences for stock.package -->
+        <record id="seq_package" model="ir.sequence">
             <field name="name">Packages</field>
-            <field name="code">stock.quant.package</field>
+            <field name="code">stock.package</field>
             <field name="prefix">PACK</field>
             <field name="padding">7</field>
             <field name="company_id" eval="False"/>
