+++ addons/stock/models/__init__.py
@@ -23,4 +23,5 @@ from . import product
 from . import product_catalog_mixin
 from . import stock_package_level
 from . import stock_package_type
+from . import stock_package
 from . import stock_storage_category
