+++ addons/stock/models/stock_package_level.py
@@ -13,7 +13,7 @@ class StockPackage_Level(models.Model):
     _check_company_auto = True
 
     package_id = fields.Many2one(
-        'stock.quant.package', 'Package', required=True, check_company=True,
+        'stock.package', 'Package', required=True, check_company=True,
         domain="[('location_id', 'child_of', parent.location_id), '|', ('company_id', '=', False), ('company_id', '=', company_id)]")
     picking_id = fields.Many2one('stock.picking', 'Picking', check_company=True, index='btree_not_null')
     move_ids = fields.One2many('stock.move', 'package_level_id')
