+++ addons/stock/security/ir.model.access.csv
@@ -21,9 +21,9 @@ access_stock_warehouse_orderpoint,stock.warehouse.orderpoint,model_stock_warehou
 access_stock_warehouse_orderpoint_system,stock.warehouse.orderpoint system,model_stock_warehouse_orderpoint,stock.group_stock_manager,1,1,1,1
 access_stock_quant_user,stock.quant user,model_stock_quant,stock.group_stock_user,1,1,1,0
 access_stock_quant_all,stock.quant all users,model_stock_quant,base.group_user,1,0,0,0
-access_stock_quant_package_all,stock.quant.package all users,model_stock_quant_package,base.group_user,1,0,0,0
-access_stock_quant_package_stock_manager,stock.quant.package stock manager,model_stock_quant_package,stock.group_stock_manager,1,1,1,1
-access_stock_quant_package_stock_user,stock.quant.package stock user,model_stock_quant_package,stock.group_stock_user,1,1,1,1
+access_stock_package_all,stock.package all users,model_stock_package,base.group_user,1,0,0,0
+access_stock_package_stock_manager,stock.package stock manager,model_stock_package,stock.group_stock_manager,1,1,1,1
+access_stock_package_stock_user,stock.package stock user,model_stock_package,stock.group_stock_user,1,1,1,1
 access_stock_package_level_all,stock.package_level all users,model_stock_package_level,base.group_user,1,0,0,0
 access_stock_package_level_stock_manager,stock.package_level stock manager,model_stock_package_level,stock.group_stock_manager,1,1,1,1
 access_stock_package_level_stock_user,stock.package_level stock user,model_stock_package_level,stock.group_stock_user,1,1,1,1
