+++ addons/stock/tests/test_move.py
@@ -941,7 +941,7 @@ class TestStockMove(TestStockCommon):
             'name': 'Super Package Type',
         })
 
-        package = self.env['stock.quant.package'].create({'package_type_id': package_type.id})
+        package = self.env['stock.package'].create({'package_type_id': package_type.id})
 
         self.env['stock.putaway.rule'].create({
             'product_id': self.productA.id,
@@ -979,7 +979,7 @@ class TestStockMove(TestStockCommon):
             'name': 'Super Package Type',
         })
 
-        package = self.env['stock.quant.package'].create({'package_type_id': package_type.id})
+        package = self.env['stock.package'].create({'package_type_id': package_type.id})
 
         self.env['stock.putaway.rule'].create({
             'product_id': self.productA.id,
@@ -1044,7 +1044,7 @@ class TestStockMove(TestStockCommon):
         package_type = self.env['stock.package.type'].create({
             'name': 'Super Package Type',
         })
-        package = self.env['stock.quant.package'].create({
+        package = self.env['stock.package'].create({
             'package_type_id': package_type.id,
         })
 
@@ -1333,7 +1333,7 @@ class TestStockMove(TestStockCommon):
             'putaway_rule_ids': [(4, putaway.id, 0)],
         })
 
-        package = self.env['stock.quant.package'].create({
+        package = self.env['stock.package'].create({
             'name': 'package',
             'package_type_id': package_type.id,
         })
@@ -1396,7 +1396,7 @@ class TestStockMove(TestStockCommon):
         })
 
         # first package
-        package1 = self.env['stock.quant.package'].create({
+        package1 = self.env['stock.package'].create({
             'name': 'package 1',
             'package_type_id': package_type.id,
         })
@@ -1424,7 +1424,7 @@ class TestStockMove(TestStockCommon):
         self.assertEqual(package1.location_id.id, self.shelf_2.id)
 
         # second package
-        package2 = self.env['stock.quant.package'].create({
+        package2 = self.env['stock.package'].create({
             'name': 'package 2',
             'package_type_id': package_type.id,
         })
@@ -1491,7 +1491,7 @@ class TestStockMove(TestStockCommon):
         })
 
         # first package
-        package1 = self.env['stock.quant.package'].create({
+        package1 = self.env['stock.package'].create({
             'name': 'package 1',
             'package_type_id': package_type.id,
         })
@@ -1519,7 +1519,7 @@ class TestStockMove(TestStockCommon):
         self.assertEqual(package1.location_id.id, self.shelf_2.id)
 
         # second package
-        package2 = self.env['stock.quant.package'].create({
+        package2 = self.env['stock.package'].create({
             'name': 'package 2',
             'package_type_id': package_type.id,
         })
@@ -1585,7 +1585,7 @@ class TestStockMove(TestStockCommon):
         })
 
         # first package
-        package1 = self.env['stock.quant.package'].create({
+        package1 = self.env['stock.package'].create({
             'name': 'package 1',
             'package_type_id': package_type.id,
         })
@@ -1613,7 +1613,7 @@ class TestStockMove(TestStockCommon):
         self.assertEqual(package1.location_id.id, self.shelf_2.id)
 
         # second package
-        package2 = self.env['stock.quant.package'].create({
+        package2 = self.env['stock.package'].create({
             'name': 'package 2',
             'package_type_id': package_type.id,
         })
@@ -2111,7 +2111,7 @@ class TestStockMove(TestStockCommon):
         """ Check that unreserving a stock move sets the products reserved as available and
         set the state back to confirmed even if they are in a pack.
         """
-        package1 = self.env['stock.quant.package'].create({'name': 'test_unreserve_2_pack'})
+        package1 = self.env['stock.package'].create({'name': 'test_unreserve_2_pack'})
 
         # make some stock
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 150.0, package_id=package1)
@@ -3265,8 +3265,8 @@ class TestStockMove(TestStockCommon):
         """ Test that editing a stock move line linked to a packed product correctly and directly
         adapts the reservation. In this case, we edit the package to another available one.
         """
-        package1 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_3'})
-        package2 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_3'})
+        package1 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_3'})
+        package2 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_3'})
 
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, package_id=package1)
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, package_id=package2)
@@ -3339,7 +3339,7 @@ class TestStockMove(TestStockCommon):
             'name': 'lot2',
             'product_id': self.productA.id,
         })
-        package1 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_5'})
+        package1 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_5'})
 
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, lot_id=lot1, package_id=package1)
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, lot_id=lot2)
@@ -3623,8 +3623,8 @@ class TestStockMove(TestStockCommon):
         """ Test that editing a done stock move line linked to a packed product correctly and directly
         adapts the transfer. In this case, we edit the package to another available one.
         """
-        package1 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_3'})
-        package2 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_3'})
+        package1 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_3'})
+        package2 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_3'})
 
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, package_id=package1)
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, package_id=package2)
@@ -3701,7 +3701,7 @@ class TestStockMove(TestStockCommon):
             'name': 'lot2',
             'product_id': self.productA.id,
         })
-        package1 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_5'})
+        package1 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_5'})
 
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, lot_id=lot1, package_id=package1)
         self.env['stock.quant']._update_available_quantity(self.productA, self.stock_location, 1.0, lot_id=lot2)
@@ -3962,7 +3962,7 @@ class TestStockMove(TestStockCommon):
             'name': 'lot2',
             'product_id': self.product_lot.id,
         })
-        self.env['stock.quant.package'].create({'name': 'test_edit_done_move_line_12'})
+        self.env['stock.package'].create({'name': 'test_edit_done_move_line_12'})
         move1 = self.env['stock.move'].create({
             'location_id': self.supplier_location.id,
             'location_dest_id': self.stock_location.id,
@@ -3995,7 +3995,7 @@ class TestStockMove(TestStockCommon):
             'name': 'lot2',
             'product_id': self.product_lot.id,
         })
-        package1 = self.env['stock.quant.package'].create({'name': 'test_edit_reserved_move_line_5'})
+        package1 = self.env['stock.package'].create({'name': 'test_edit_reserved_move_line_5'})
 
         move1 = self.env['stock.move'].create({
             'location_id': self.supplier_location.id,
@@ -6345,7 +6345,7 @@ class TestStockMove(TestStockCommon):
             'product_id': self.product_serial.id,
         } for name in ['sn01', 'sn02', 'sn03']])
 
-        pack = self.env['stock.quant.package'].create({
+        pack = self.env['stock.package'].create({
             'name': 'Pack A',
         })
