+++ addons/stock/views/stock_quant_views.xml
@@ -244,128 +244,6 @@ No Stock On Hand. This analysis gives you an overview of the current stock level
         </field>
     </record>
 
-    <record id="quant_package_search_view" model="ir.ui.view">
-        <field name="name">stock.quant.package.search</field>
-        <field name="model">stock.quant.package</field>
-        <field eval="10" name="priority"/>
-        <field name="arch" type="xml">
-            <search string="Package">
-                <field name="name" string="Package Name"/>
-                <field name="location_id"/>
-                <field name="package_type_id"/>
-                <filter string="In internal locations" name="internal" domain="['|', ('location_id.usage', '=', 'internal'), ('location_id', '=', False)]" groups="stock.group_stock_multi_locations"/>
-                <group>
-                   <filter string='Location' name="location" domain="[]" context="{'group_by' : 'location_id'}" groups="stock.group_stock_multi_locations"/>
-                   <filter string='Package Type' name="package_type" domain="[]" context="{'group_by' : 'package_type_id'}"/>
-               </group>
-            </search>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="view_quant_package_form">
-        <field name="name">stock.quant.package.form</field>
-        <field name="model">stock.quant.package</field>
-        <field eval="10" name="priority"/>
-        <field name="arch" type="xml">
-            <form string="Package">
-                <header>
-                    <button name="unpack" string="Unpack" type="object"/>
-                </header>
-                <sheet>
-                    <div class="oe_button_box" name="button_box">
-                        <button class="oe_stat_button" name="action_view_picking"
-                        string="Package Transfers" type="object"
-                        widget="statinfo" icon="oi-arrows-v"/>
-                    </div>
-                    <div class="oe_title">
-                        <label for="name" string="Package Reference"/>
-                        <h1><field name="name" class="oe_inline" placeholder="e.g. PACK0000007"/></h1>
-                    </div>
-                    <group>
-                        <group>
-                            <field name="package_type_id"/>
-                            <field name='company_id' groups="base.group_multi_company"/>
-                            <field name='owner_id' groups="stock.group_tracking_owner"/>
-                            <field name="location_id" options="{'no_create': True}"/>
-                        </group>
-                        <group>
-                            <field name="pack_date"/>
-                        </group>
-                    </group>
-                    <separator string="Content"/>
-                    <field name="quant_ids">
-                        <list>
-                            <field name="product_id"/>
-                            <field name="lot_id" groups="stock.group_production_lot"/>
-                            <field name="quantity"/>
-                            <field name="product_uom_id" groups="uom.group_uom" widget="many2one_uom"/>
-                        </list>
-                        <kanban class="o_kanban_mobile">
-                            <templates>
-                                <t t-name="card">
-                                    <field name="product_id" class="fw-bold"/>
-                                    <field name="lot_id"/>
-                                    <div class="d-flex">
-                                        <field name="quantity"/>
-                                        <field class="mx-2" name="product_uom_id" groups="uom.group_uom"/>
-                                    </div>
-                                </t>
-                            </templates>
-                        </kanban>
-                    </field>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="view_quant_package_tree">
-        <field name="name">stock.quant.package.list</field>
-        <field name="model">stock.quant.package</field>
-        <field eval="10" name="priority"/>
-        <field name="arch" type="xml">
-            <list string="Package" sample="1">
-                <field name="display_name"/>
-                <field name="package_type_id"/>
-                <field name="location_id" options="{'no_create': True}"/>
-                <field name="company_id" groups="base.group_multi_company"/>
-            </list>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="view_quant_package_kanban">
-        <field name="name">stock.quant.package.kanban</field>
-        <field name="model">stock.quant.package</field>
-        <field name="arch" type="xml">
-            <kanban sample="1" group_create="0">
-                <templates>
-                    <t t-name="card" class="row g-0">
-                        <field name="name" class="col-6 fw-bold fs-5"/>
-                        <field name="package_type_id" class="col-6"/>
-                    </t>
-                </templates>
-            </kanban>
-        </field>
-    </record>
-
-    <record model="ir.actions.act_window" id="action_package_view">
-        <field name="name">Packages</field>
-        <field name="res_model">stock.quant.package</field>
-        <field name="view_mode">kanban,list,form</field>
-        <field name="context">{
-            'search_default_location': True,
-            'search_default_internal': True,
-            }
-        </field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-                Create a new package
-            </p><p>
-                Packages are usually created via transfers (during pack operation) and can contain different products.
-                Once created, the whole package can be moved at once, or products can be unpacked and moved as single units again.
-            </p>
-        </field>
-    </record>
-
     <record id="duplicated_sn_warning" model="ir.ui.view">
         <field name="name">stock.quant.duplicated.sn.warning</field>
         <field name="priority">1000</field>
@@ -492,7 +370,6 @@ in this location. That leads to a negative stock.
     </record>
 
     <menuitem id="menu_action_inventory_tree" name="Physical Inventory" parent="menu_stock_adjustments" sequence="10" action="action_view_inventory_tree"/>
-    <menuitem id="menu_package" name="Packages" parent="menu_stock_inventory_control" action="action_package_view" groups="stock.group_tracking_lot" sequence="102"/>
     <menuitem id="menu_valuation" name="Locations"
               parent="stock.menu_warehouse_report" sequence="150"
               action="action_view_quants" groups="stock.group_stock_multi_locations,stock.group_tracking_owner,base.group_no_one"/>
