+++ addons/stock_delivery/models/__init__.py
@@ -6,4 +6,4 @@ from . import sale_order
 from . import stock_move
 from . import stock_package_type
 from . import stock_picking
-from . import stock_quant_package
+from . import stock_package
